write to current world instead of new world and remove multithreaded implementation
The genesheep algorithm actually doesn't use 2 worlds and instead write
directly back to the current world. Because of this the multithreaded
implemenation needs a lot of rework so removed it for now.

Further optimized the singlethreaded implementation and shufle the
to_check at the start to reduce bias in the result.